Allow multiple module use and change the way information are stored in the SED
The modules can now be named with any prefix using a dot (e.g.
dh2002_ir.1). That means that a 'pristine' module can't have a dot in
it's name.

The add_component SED method was removed. To add components to a sed,
add_info, add_module (NEW) and add_contribution (NEW) must be used.

The modules now add more information in the sed info dictionary, even
the parametres that can be used for the statistical analysis. In
particular, mh2005_sfh add the SFR and the average SFR.

The sfr variable was renamed to sfh.