Correct the Maraston 2005 module to use SFH
The Marasont 2005 module is modified to use the output of a star
formation history module. The SFH is now a complete SFR vs time tuple
(taken from the SED information dictionary) and the age of the galaxy is
defined by this SFH (the maximum time).
The Maraston 2005 data class is also changed to make the convolution in
such SFH.