Cast numeric parameters to their correct type
To avoid problems that might occur when, for instance, using an integer
instead of a float, we now cast the module parameters to their expected
type. This will also have the advantage to raise an error when a
parameter can't be interpreted with its expected type.

Also, use integer for age parameters as we are now using Myr as unit.