Initialise module parameters at instantiation
In order to speed-up module re-use, get the module needs (ssp,
filters...) out of pcigale database at module instantiation instead of
doing this everytime "process" is called.

To do that:
- we must give the module parameters (and proceed with their checking)
at
  module instantiation.
- we must add the possibility for each module to have its own
  initialisation code.

In addition, to create the configuration file, we must be able to load
the modules without any parameter to query them for their parameter
list.

The bc03, dh2002 and m2005 modules were changed to use this new
behavior.