Get rid of process / _process separation
The process method was used to complete and check the parameter
dictionary before launching the _process one defined in each module. Now
that the parameters are dealed with at instantiation, this separation is
no more useful.

Also, modify the attenuation modules to read the filter in the database
at instantiation.