Use & to enter a list as one parameter value
When one value for a parameter is a list (for instance the list of
attenuation keys for re-emission modules) we had to use the "eval [["a",
"b"]]" notation. Convert these module to use a single string with items
separated by a & ("a & b").

WARNING we must still change the way parameters are evaluated not to
return everytime a list.