Postfix the key instead of prefixing
Instead of prefixing the key in the SED with the module name, we postfix
them with the same postfix as the module name if any. This makes easier
to inter-change modules adding the same information (for instance
attenuation modules adding the attenuation value).
That implies that when using several modules adding the same
information, one must postfix the name of each.