Make SFH normalisation optional in sfhfromfile
In the sfhfromfile module, make the SFH normalisation to 1 solar mass
produced at the given age optional. When using pcigale to simulate SEDs,
one may not want this normalisation to happend.

Such SEDs should not be stored in the database (but SEDs made using
sfhfromfile should not anyway be stored in the database). We will have
to make the SED storing optional in the warehouse.