Change how redshit and fluxes are managed
Now the redshift is an attribute of the SED.

The SED.compute_fnu() no longuer takes redshift and apply_redshift
arguments as the SED is supposed to be redshifted.

The SED.lambda_fnu() method was converted to the SED.fnu property
returning th Fν flux density at each wavelength of the wavelength grid.