New pseudo filters to compute line fluxes
Addition of pseudo filters used to compute de fluxes in some lines.
They are composed of a positive part (normalised to 1) above the line
and two negative part (of the same total area) on each side used to
remove the continuum.

The PSEUDO_D4000 pseudo filter measures the difference (not the ratio)
of the 4000 Å break.