Make the filter list optional in the param module
Using the param module, if the list of filters is empty, no flux will be
computed and the module won't complain about '' missing filter.  Also,
switch the default value for this list of filters to the empty string,
because None means that there is no default value which makes the module
not usable without the list for filters parameters.

Closes #5