Homogenise attenuation modules outputs
In the attenuation modules, remove the attenuation.ebvs_main and
attenuation.ebvs_young parameter that are duplicates of
attenuation.E_BVs.stellar.old and attenuation.E_BVs.stellar.young.

Homogenise the outputs of the two modules and add the Av/E_BVs old
factors to the information dictionary so that their analysis can be
done.