Convert input table to float
When reading a table from a file, convert all the integer columns to float
(except the 'id' column).  These columns are fluxes or error on fluxes and thus
should be treated as float even when they contain only integer.  This solves the
problem of trying to affect NaN value to an integer column.

Fixes #6